git-svn-id: svn://123.57.173.36/meijia-parent@1578 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/record/RecordAssetUseMapper.java
@@ -10,7 +10,7 @@ public interface RecordAssetUseMapper {
 
     int insert(RecordAssetUse record);
 
-    int insertSelective(RecordAssetUse record);
+    Long insertSelective(RecordAssetUse record);
 
     RecordAssetUse selectByPrimaryKey(Long id);
